python-almost_a_circle/models/base.py
#!/usr/bin/python3
"""Module for Base class."""
import json


class Base:
    """Base class for other classes in the project."""
    __nb_objects = 0

    def __init__(self, id=None):
        """Initializes the Base instance.

        Args:
            id (int): The id of the instance.
        """
        if id is not None:
            self.id = id
        else:
            Base.__nb_objects += 1
            self.id = Base.__nb_objects

    @staticmethod
    def to_json_string(list_dictionaries):
        """Returns the JSON string representation of list_dictionaries.

        Args:
            list_dictionaries (list): A list of dictionaries.
        """
        if list_dictionaries is None or len(list_dictionaries) == 0:
            return "[]"
        return json.dumps(list_dictionaries)

    @staticmethod
    def from_json_string(json_string):
        """Returns the list of the JSON string representation json_string.

        Args:
            json_string (str): A string representing a list of dictionaries.
        """
        if json_string is None or len(json_string) == 0:
            return []
        return json.loads(json_string)

    @classmethod
    def create(cls, **dictionary):
        """Returns an instance with all attributes already set.

        Args:
            **dictionary (dict): Double pointer to a dictionary.
        """
        # Dummy instance creation will depend on the specific class
        # and should be implemented in child classes.
        pass

    @classmethod
    def load_from_file(cls):
        """Returns a list of instances."""
        # This method should be implemented in child classes.
        pass
